# docs/api/openapi_spec.yaml
openapi: 3.0.1
info:
  title: Fantasy Football API
  version: v1
  description: >
    API for onboarding and querying fantasy football league data.
    Supports ESPN and Sleeper as data sources. League data is fetched
    from external APIs, transformed, and stored in DynamoDB for fast retrieval.
  license:
    name: MIT
paths:
...
  /leagues/{leagueId}:
    get:
      summary: Get league metadata
      description: >
        Checks whether a league has been onboarded. If the league exists,
        returns the full league lookup item including the canonical league ID
        and the list of seasons the league has been active for. Returns 404
        if the league has not been onboarded yet.
      operationId: getLeague
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ESPN or Sleeper league ID
          schema:
            type: string
            pattern: '^\d+$'
            example: "12345678"
        - name: platform
          in: query
          required: true
          description: The name of the platform the fantasy football league is on
          schema:
            type: string
            enum: [ESPN, SLEEPER]
            example: "ESPN"
      responses:
        "200":
          $ref: "#/components/responses/LeagueFoundResponse"
        "404":
          $ref: "#/components/responses/LeagueNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)

    delete:
      summary: Delete league
      description: >
        Deletes an onboarded league and its associated data from DynamoDB and S3.
        Also decrements the total onboarded league count.
      operationId: deleteLeague
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ESPN or Sleeper league ID
          schema:
            type: string
            pattern: '^\d+$'
            example: "12345678"
        - name: platform
          in: query
          required: true
          description: The name of the platform the fantasy football league is on
          schema:
            type: string
            enum: [ESPN, SLEEPER]
            example: "ESPN"
      responses:
        "200":
          $ref: "#/components/responses/DeleteLeagueResponse"
        "404":
          $ref: "#/components/responses/LeagueNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
  /leagues:
    post:
      summary: Onboard a new league or refresh an existing league's data
      description: >
        Triggers the onboarding Lambda which fetches all historical season data
        from the specified platform (ESPN or Sleeper), transforms it via DuckDB,
        and writes precomputed views to DynamoDB. The requestType query parameter
        determines whether this is a new onboarding ("ONBOARD") or a refresh of an
        existing league's data ("REFRESH").
      operationId: onboardLeague
      parameters:
        - name: requestType
          in: query
          required: true
          description: > 
            The type of onboarding request being made. Can be "ONBOARD" for new leagues
            or "REFRESH" to refresh data for an existing league.
          schema:
            type: string
            enum: [ONBOARD, REFRESH]
            example: "ONBOARD"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardRequest"
      responses:
        "201":
          $ref: "#/components/responses/OnboardSuccessResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
  /leagues/{leagueId}/query:
    get:
      summary: Query a precomputed league view
      description: >
        Returns a precomputed data view for the specified league. The queryType
        parameter maps directly to the DynamoDB sort key (e.g. WEEKLY_SCORES#2024).
      operationId: queryLeague
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ESPN or Sleeper league ID
          schema:
            type: string
            pattern: '^\d+$'
            example: "12345678"
        - name: platform
          in: query
          required: true
          description: The name of the platform the fantasy football league is on
          schema:
            type: string
            enum: [ESPN, SLEEPER]
            example: "ESPN"
        - name: queryType
          in: query
          required: true
          description: >
            The precomputed view to retrieve. Season-scoped views require
            a season suffix (e.g. SEASON_STANDINGS#2024). Cross-season views
            such as ALL_TIME_STANDINGS do not require a suffix.
          schema:
            type: string
            # TODO: Fill this out as the list of views evolves
            enum:
              - TEAMS
              - MATCHUPS
              - SEASON_STANDINGS
            example: "SEASON_STANDINGS"
      responses:
        "200":
          $ref: "#/components/responses/QuerySuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/LeagueNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
  /leagues/{leagueId}/refresh_status:
    get:
      summary: Check onboarding/refresh status for a league
      description: >
        Returns the current status of an onboarding or refresh operation for the specified league.
        Used for polling after triggering onboarding or a refresh to know when the data is ready.
      operationId: checkRefreshStatus
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ESPN or Sleeper league ID
          schema:
            type: string
            pattern: '^\d+$'
            example: "12345678"
        - name: platform
          in: query
          required: true
          description: The name of the platform the fantasy football league is on
          schema:
            type: string
            enum: [ESPN, SLEEPER]
            example: "ESPN"
        - name: refreshOperation
          in: query
          required: true
          description: The type of refresh ("ONBOARD" or "REFRESH") to check the status of.
          schema:
            type: string
            enum: [ONBOARD, REFRESH]
            example: "ONBOARD"
      responses:
        "200":
          ref: "#/components/responses/LeagueStatusSuccessResponse"
        "404":
          $ref: "#/components/responses/LeagueNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
  /leagues/migrate/new_users:
    post:
      summary: Get users from target league for migration
      description: >
        Fetches users from the target league (ESPN or Sleeper) for migration purposes.
        For ESPN leagues, requires season, s2, and swid credentials. For Sleeper leagues,
        only the league ID is required.
      operationId: getMigrateUsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MigrateUsersRequest"
      responses:
        "200":
          $ref: "#/components/responses/MigrateUsersSuccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
components:
  securitySchemes:
    ClerkJWT:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: jwt
      x-amazon-apigateway-authorizer:
        identitySource: $request.header.Authorization
        jwtConfiguration:
          issuer: "${clerk_issuer_url}"
          audience:
            - "${clerk_jwt_audience}"
        type: jwt
  schemas:
...
    OnboardRequest:
      type: object
      required: [leagueId, platform]
      properties:
        leagueId:
          type: string
          description: The ESPN or Sleeper league ID
          example: "12345678"
        platform:
          type: string
          enum: [ESPN, SLEEPER]
          description: The fantasy platform this league belongs to
          example: "ESPN"
        season:
          type: string
          description: >
            The most recent season the league was active for. Required for
            ESPN leagues only.
          example: "2025"
        s2:
          type: string
          description: >
            ESPN espn_s2 cookie value. Required for private ESPN leagues only.
            This value is only passed to the application for onboarding and is
            never logged or stored in our database.
          example: "AEBxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
        swid:
          type: string
          description: >
            ESPN SWID cookie value. Required for private ESPN leagues only.
            This value is only passed to the application for onboarding and is
            never logged or stored in our database.
          example: "{XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX}"
...
  responses:
...
    LeagueFoundResponse:
      description: League found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/APIResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/LeagueFoundData"
          example:
            detail: "Found league"
            data:
              canonical_league_id: "1d97a0ac-1a3c-4dd7-a4a0-a6f74035d076"
              seasons: ["2023", "2024", "2025"]
    DeleteLeagueResponse:
      description: League successfully deleted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            detail: "Successfully deleted league"
    OnboardSuccessResponse:
      description: Onboarding successfully triggered
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            detail: "Successfully triggered onboarding"
...
    LeagueNotFoundError:
      description: League not found — onboarding required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "League not found"
            message: "No ESPN league with ID 12345678 has been onboarded. Use POST /leagues to onboard."
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Internal server error"
            message: "An unexpected error occurred. Please try again later."